requirements.txt
aiogram>=3.7,<4
SQLAlchemy>=2.0,<3
aiosqlite>=0.20,<1
python-dotenv>=1.0,<2
pytest>=8.0,<9
